chore: split conten component
