+++ src/components/organisms/Dashboard/Header/Header.tsx
@@ -1,0 +1,11 @@
+import {TitleBar} from '@monokle/components';
+
+import * as S from './Header.styled';
+
+export const Header = () => {
+  return (
+    <S.Container>
+      <TitleBar type="secondary" title="Overview" actions={<span>Selection</span>} />
+    </S.Container>
+  );
+};
