+++ src/components/organisms/Dashboard/Tableview/Tableview.tsx
@@ -1,0 +1,3 @@
+export const Tableview = () => {
+  return <div>Tableview</div>;
+};
